+++ R/rbindFill.R
@@ -1,6 +1,7 @@
 #' @title Combine R Objects by Row
 #'
 #' @description
+#'
 #' This function combines instances of `matrix`, `data.frame` or `DataFrame`
 #' objects into a single instance adding eventually missing columns (filling
 #' them with `NA`s).
@@ -8,18 +9,23 @@
 #' @param ... 2 or more: `matrix`, `data.frame` or `DataFrame`.
 #'
 #' @return Depending on the input a single `matrix`, `data.frame` or
-#' `DataFrame`.
+#'     `DataFrame`.
 #'
 #' @note
-#' `rbindFill` might not work if one of the columns contains S4 classes.
+#'
+#' `rbindFill()` might not work if one of the columns contains S4 classes.
 #'
 #'
 #' @author Johannes Rainer, Sebastian Gibb
 #'
 #' @family helper functions for developers
+#'
 #' @importMethodsFrom S4Vectors cbind nrow rownames colnames
+#'
 #' @importFrom methods as
+#'
 #' @export
+#'
 #' @examples
 #' ## Combine matrices
 #' a <- matrix(1:9, nrow = 3, ncol = 3)
@@ -28,6 +34,16 @@
 #' colnames(b) <- c("b", "a", "d", "e")
 #' rbindFill(a, b)
 #' rbindFill(b, a, b)
+#'
+#' ## Combine data.frame
+#' d <- data.frame(a = 1:4, z = rep(TRUE, 4))
+#' g <- data.frame(b = 1:3, z = rep(FALSE, 3))
+#' rbindFill(d, g)
+#'
+#' ## Combine matrix and data.frames
+#' res <- rbindFill(a, d)
+#' res
+#' class(res)
 rbindFill <- function(...) {
     l <- list(...)
 
@@ -48,6 +64,8 @@ rbindFill <- function(...) {
     ## convert matrix to data.frame for easier and equal subsetting and class
     ## determination
     isMatrix <- as.logical(cls["matrix",])
+    ## if (all(isMatrix))
+    ##     return(.rbind_fill_matrix(l))
     l[isMatrix] <- lapply(l[isMatrix], as.data.frame)
 
     allcl <- unlist(
@@ -74,3 +92,35 @@ rbindFill <- function(...) {
 
 ## helper function to allow lapply(..., as, ...) in rbindFill
 setAs("logical", "factor", function(from, to) factor(from))
+
+#' @description
+#'
+#' Combines provided matrices into a single matrix eventually adding missing
+#' columns (filled with `NA`). After adding missing columns, the matrices
+#' are combined with `base::rbind` hence eventually changing the type of the
+#' matrices.
+#'
+#' @param l `list` of `matrix`
+#'
+#' @return `matrix` with combined matrices from `l`
+#'
+#' @author Johannes Rainer
+#'
+#' @noRd
+.rbind_fill_matrix <- function(l) {
+    cnames <- unique(unlist(lapply(l, colnames)))
+    do.call(rbind, lapply(l, function(z) {
+        cnz <- colnames(z)
+        if (length(cnz) != ncol(z))
+            stop("All matrices need to have column names", call. = FALSE)
+        miss <- setdiff(cnames, cnz)
+        if (length(miss)) {
+            z <- cbind(z, matrix(NA, ncol = length(miss),
+                                 nrow = nrow(z)))
+            colnames(z) <- c(cnz, miss)
+        }
+        if (!identical(unname(colnames(z)), unname(cnames)))
+            z <- z[, cnames]
+        z
+    }))
+}