+++ src/which.c
@@ -14,8 +14,8 @@
  * \note Modified from src/main/summary.c:do_which_min()
  */
 SEXP C_which_first(SEXP x) {
-    if (!isLogical(x))
-        error("argument to 'which.first' is not logical");
+    if (!Rf_isLogical(x))
+        Rf_error("argument to 'which.first' is not logical");
 
     const R_len_t n = XLENGTH(x);
     R_len_t i = 0, j = -1;
@@ -28,15 +28,15 @@ SEXP C_which_first(SEXP x) {
     }
 
     i = (j != -1);
-    SEXP r = PROTECT(allocVector(INTSXP, i ? 1 : 0));
+    SEXP r = PROTECT(Rf_allocVector(INTSXP, i ? 1 : 0));
     if (i) {
         INTEGER(r)[0] = j + 1;
-        if (getAttrib(x, R_NamesSymbol) != R_NilValue) { /* preserve names */
+        if (Rf_getAttrib(x, R_NamesSymbol) != R_NilValue) { /* preserve names */
 	        SEXP name;
 	        PROTECT(
-                name = ScalarString(
-                    STRING_ELT(getAttrib(x, R_NamesSymbol), j)));
-	        setAttrib(r, R_NamesSymbol, name);
+                name = Rf_ScalarString(
+                    STRING_ELT(Rf_getAttrib(x, R_NamesSymbol), j)));
+	        Rf_setAttrib(r, R_NamesSymbol, name);
 	        UNPROTECT(1);
         }
 	}
@@ -54,8 +54,8 @@ SEXP C_which_first(SEXP x) {
  * \note Modified from src/main/summary.c:do_which_min()
  */
 SEXP C_which_last(SEXP x) {
-    if (!isLogical(x))
-        error("argument to 'which.last' is not logical");
+    if (!Rf_isLogical(x))
+        Rf_error("argument to 'which.last' is not logical");
 
     const R_len_t n = XLENGTH(x);
     R_len_t i = n - 1, j = -1;
@@ -68,15 +68,15 @@ SEXP C_which_last(SEXP x) {
     }
 
     i = (j != -1);
-    SEXP r = PROTECT(allocVector(INTSXP, i ? 1 : 0));
+    SEXP r = PROTECT(Rf_allocVector(INTSXP, i ? 1 : 0));
     if (i) {
         INTEGER(r)[0] = j + 1;
-        if (getAttrib(x, R_NamesSymbol) != R_NilValue) { /* preserve names */
+        if (Rf_getAttrib(x, R_NamesSymbol) != R_NilValue) { /* preserve names */
 	        SEXP name;
 	        PROTECT(
-                name = ScalarString(
-                    STRING_ELT(getAttrib(x, R_NamesSymbol), j)));
-	        setAttrib(r, R_NamesSymbol, name);
+                name = Rf_ScalarString(
+                    STRING_ELT(Rf_getAttrib(x, R_NamesSymbol), j)));
+	        Rf_setAttrib(r, R_NamesSymbol, name);
 	        UNPROTECT(1);
         }
     }